% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-prop2-test.R
\name{n_prop2_test}
\alias{n_prop2_test}
\title{Sample Size for Two Sample Test Proportion}
\usage{
n_prop2_test(p1, p2, one_sided = FALSE)
}
\arguments{
\item{p1}{Proportion appearing in one group.}

\item{p2}{Proportion appearing in other group.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}
}
\value{
A integer of the required number of individuals.
}
\description{
Uses \code{stats::power.prop.test()} to calculate the number of samples that
must be included in a study
to achieve a statistical power of 0.8 at a significance level of 0.05
for a test of a difference in the proportion of samples
appearing in one of two groups.
}
\examples{
n_prop2_test(0.9, 0.1)
n_prop2_test(0.1, 0.9)
n_prop2_test(0.01, 0.09)
n_prop2_test(0.45, 0.55)
n_prop2_test(0.005, 0.0075)
}
