% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2svalue.R
\name{p2svalue}
\alias{p2svalue}
\title{Pvalue to Svalue}
\usage{
p2svalue(x)
}
\arguments{
\item{x}{A numeric vector of pvalues.}
}
\value{
A numeric vector of the equivalent svalues.
}
\description{
Pvalue to Svalue
}
\examples{
p2svalue(c(1, 0.5, 0.25, 0.05, 0.01, 0.001, 0.0001))
}
