% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-prop2-test.R
\name{p_prop2_test}
\alias{p_prop2_test}
\title{P-value for Exact Two Sample Test Proportions}
\usage{
p_prop2_test(n1, n2, one_sided = FALSE)
}
\arguments{
\item{n1}{Number of samples in first group.}

\item{n2}{Number of samples in second group.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}
}
\value{
A numeric scalar of the p-value.
}
\description{
Uses \code{stats::binom.test()} to calculate the exact p-value for a
a two sample test of a difference in the probability of being in n1 vs n2.
}
\examples{
p_prop2_test(6, 0)
p_prop2_test(3, 3)
p_prop2_test(0, 6)
p_prop2_test(106, 100)
p_prop2_test(106, 100, one_sided = TRUE)
p_prop2_test(100, 106, one_sided = TRUE)
}
