% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-propn-test.R
\name{p_props_test}
\alias{p_props_test}
\title{P-value for Exact Multiple Sample Test Proportions}
\usage{
p_props_test(ns)
}
\arguments{
\item{ns}{A vector of the number of samples in each group.}
}
\value{
A numeric scalar of the p-value.
}
\description{
Uses \code{stats::prop.test()} to calculate the exact p-value for a
multiple sample test of a difference in the probability of being in the groups.
}
\examples{
p_props_test(c(106, 100))
p_props_test(c(106, 100, 106))
p_props_test(c(106, 80))
p_props_test(c(106, 80, 106))
}
