% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dbeta.R
\name{plot_dbeta}
\alias{plot_dbeta}
\title{Plot Density for Beta Distribution}
\usage{
plot_dbeta(alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{A numeric vector of the first shape parameter for the beta distribution.}

\item{beta}{A numeric vector of the second shape parameter for the beta distribution.}
}
\description{
Plot Density for Beta Distribution
}
\examples{
plot_dbeta()
plot_dbeta(10, 10)
}
