% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-prop-test.R
\name{power_prop_test}
\alias{power_prop_test}
\title{Power Calculations for Two Sample Test Proportions}
\usage{
power_prop_test(p1, p2)
}
\arguments{
\item{p1}{Proportion appearing in one group.}

\item{p2}{Proportion appearing in other group.}
}
\value{
A integer of the required number of individuals.
}
\description{
Uses \code{power.prop.test()} to calculate the number of individuals that
must be included in a study
to achieve a statistical power of 0.8 at a significance level of 0.05
for a one-sided test of a difference in the proportion of individuals
appearing in one of two groups.
}
\examples{
power_prop_test(0.9, 0.1)
power_prop_test(0.1, 0.9)
power_prop_test(0.01, 0.09)
power_prop_test(0.45, 0.55)
power_prop_test(0.005, 0.0075)
}
