% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-ci.R
\name{prop2_ci}
\alias{prop2_ci}
\title{Calculate Estimate and Confidence Limits for Two Sample Test Proportion}
\usage{
prop2_ci(n1, n2)
}
\arguments{
\item{n1}{Number of samples in first group.}

\item{n2}{Number of samples in second group.}
}
\value{
A data frame of the estimated proportion in the first group with lower and upper 95\% confidence limits.
}
\description{
Uses \code{\link[stats:binom.test]{stats::binom.test}} to calculate the confidence intervals.
}
\examples{
prop2_ci(6, 0)
prop2_ci(3, 3)
prop2_ci(0, 6)
prop2_ci(106, 100)
prop2_ci(100, 106)
}
