% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-power-analysis.R
\name{prop2_power_analysis}
\alias{prop2_power_analysis}
\title{Power for Two Sample Test Proportion}
\usage{
prop2_power_analysis(p1, n, one_sided = FALSE, nsims = 1000)
}
\arguments{
\item{p1}{Proportion appearing in first group.}

\item{n}{Total number of samples.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}

\item{nsims}{A whole number of the number of simulations to perform.}
}
\value{
A real scalar of the power.
}
\description{
Power for Two Sample Test Proportion
}
\examples{
set.seed(99)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.9, 10)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.1, 10)
set.seed(99)
prop2_power_analysis(0.45, 500)
}
