% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-samplesize.R
\name{prop2_samplesize}
\alias{prop2_samplesize}
\title{Sample Size for Two Sample Test Proportion}
\usage{
prop2_samplesize(p1, p2 = 1 - p1, one_sided = FALSE)
}
\arguments{
\item{p1}{Proportion appearing in one group.}

\item{p2}{Proportion appearing in other group.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}
}
\value{
A integer of the required number of individuals.
}
\description{
Uses \code{stats::power.prop.test()} to calculate the number of samples that
must be included in a study
to achieve a statistical power of 0.8 at a significance level of 0.05
for a test of a difference in the proportion of samples
appearing in one of two groups.
}
\examples{
prop2_samplesize(0.9)
prop2_samplesize(0.1)
prop2_samplesize(0.01, 0.09)
prop2_samplesize(0.45)
prop2_samplesize(0.005, 0.0075)
prop2_samplesize(0.005, 0.0075, one_sided = TRUE)
}
