% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-sim.R
\name{prop2_sim}
\alias{prop2_sim}
\title{Simulate Data for Two Sample Test Proportion}
\usage{
prop2_sim(p1, p2, n)
}
\arguments{
\item{p1}{Proportion appearing in first group.}

\item{p2}{Proportion appearing in second group.}

\item{n}{Total number of samples.}
}
\value{
A real vector of the number in the first and second group.
}
\description{
Simulate Data for Two Sample Test Proportion
}
\examples{
prop2_sim(0.9, 0.1, 10)
prop2_sim(0.9, 0.1, 10)
prop2_sim(0.1, 0.9, 10)
prop2_sim(0.01, 0.09, 10)
prop2_sim(0.45, 0.55, 10)
prop2_sim(0.005, 0.0075, 1000)
}
