% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn-ci.R
\name{propn_ci}
\alias{propn_ci}
\title{P-value for Multiple Sample Test Proportions}
\usage{
propn_ci(ns)
}
\arguments{
\item{ns}{A vector of the number of samples in each group.}
}
\value{
A data set of the estimate and lower and upper 95\% confidence limits.
}
\description{
P-value for Multiple Sample Test Proportions
}
\examples{
propn_ci(c(106, 100))
propn_ci(c(106, 100, 90))
propn_ci(c(106, 80))
propn_ci(c(106, 80, 106))
}
