% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn-pvalue.R
\name{propn_pvalue}
\alias{propn_pvalue}
\title{P-value for Multiple Sample Test Proportions}
\usage{
propn_pvalue(ns, p = NULL)
}
\arguments{
\item{ns}{A vector of the number of samples in each group.}

\item{p}{A vector of the null probabilities of succes If NULL the
probabilities are assumed to be equal.}
}
\value{
A numeric scalar of the p-value.
}
\description{
P-value for Multiple Sample Test Proportions
}
\examples{
propn_pvalue(c(106, 100))
propn_pvalue(c(106, 100, 90))
propn_pvalue(c(106, 80))
propn_pvalue(c(106, 80, 106))
propn_pvalue(c(60, 40))
propn_pvalue(c(60, 40), p = c(0.6, 0.4))
propn_pvalue(c(60, 40), p = c(0.4, 0.6))
}
