% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate2-power-analysis.R
\name{rate2_power_analysis}
\alias{rate2_power_analysis}
\title{Power Analysis for Rates}
\usage{
rate2_power_analysis(p1, n, p2 = NULL, alternative = "two.sided", nsims = 1000)
}
\arguments{
\item{p1}{A numeric scalar of the proportion of individuals expected to return from the first group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}

\item{p2}{A numeric scalar of the proportion of individuals expected to return from the second group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{nsims}{Number of simulations to perform.}
}
\value{
A real scalar of the power.
}
\description{
If p2 is NULL estimates the power for detecting whether p1 is different from 0.5.
Other estimates the power for detecting a difference between the two groups.
}
\examples{
rate2_power_analysis(0.9, 10, nsims = 100)
rate2_power(0.005, p2 = 0.0075, 1000, alternative = "greater")
}
