% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate2-power-analysis-bayesian.R
\name{rate2_power_analysis_bayesian}
\alias{rate2_power_analysis_bayesian}
\title{Power Analysis for Rates using Bayesian Model}
\usage{
rate2_power_analysis_bayesian(
  p1,
  n,
  p2 = NULL,
  alpha1 = 1,
  beta1 = 1,
  alpha2 = alpha1,
  beta2 = beta1,
  alternative = "two.sided",
  nsims = 1000,
  niter = 10000
)
}
\arguments{
\item{p1}{A numeric scalar of the proportion of individuals expected to return from the first group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}

\item{p2}{A numeric scalar of the proportion of individuals expected to return from the second group.}

\item{alpha1}{A numeric vector of the first shape parameter for the beta distribution for the first group.}

\item{beta1}{A numeric vector of the second shape parameter for the beta distribution for the first group.}

\item{alpha2}{A numeric vector of the first shape parameter for the beta distribution for the second group.}

\item{beta2}{A numeric vector of the second shape parameter for the beta distribution for the second group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{nsims}{Number of simulations to perform.}

\item{niter}{The number of iterations per chain for MCMC sampling.}
}
\value{
A real scalar of the power.
}
\description{
Uses \code{\link[=rate_sim]{rate_sim()}} and \code{\link[=rate_effect]{rate_effect()}} to simulate the statistical
power (probability of a p-value < 0.05).
If p2 is NULL estimates the power for detecting whether p1 is different from 0.5.
Otherwise estimates the power for detecting a difference between the two groups.
}
\examples{
rate2_power_analysis_bayesian(0.9, 10, nsims = 100)
rate2_power_analysis_bayesian(0.005, p2 = 0.0075, n = 1000, alternative = "greater", nsims = 100)
}
\seealso{
\code{\link[=rate2_power]{rate2_power()}} and \code{\link[=rate_effect]{rate_effect()}}
}
