% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate2-samplesize.R
\name{rate2_samplesize}
\alias{rate2_samplesize}
\title{Calculate Sample Size for Two Sample Test Proportion}
\usage{
rate2_samplesize(p1, p2 = 1 - p1, alternative = "two.sided")
}
\arguments{
\item{p1}{A numeric scalar of the proportion of individuals expected to return from the first group.}

\item{p2}{A numeric scalar of the proportion of individuals expected to return from the second group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}
}
\value{
A integer of the required total sample size.
}
\description{
Uses \code{\link[stats:power.prop.test]{stats::power.prop.test()}} to calculate the total number of samples that
must be included in a study
to achieve a statistical power of 0.8 at a significance level of 0.05
for a test of a difference in the proportion of samples
appearing in two groups.
}
\details{
It is worth noting that \code{n} is the total number of samples required in the study,
unlike with \code{\link[stats:power.prop.test]{stats::power.prop.test()}} where it is the total number of
samples in each group.
}
\examples{
rate2_samplesize(0.9)
rate2_samplesize(0.1)
rate2_samplesize(0.01, 0.09)
rate2_samplesize(0.45)
rate2_samplesize(0.005, 0.0075)
rate2_samplesize(0.005, 0.0075, alternative = "less")
rate2_samplesize(0.005, 0.0075, alternative = "greater")
}
\seealso{
\code{\link[=rate2_power]{rate2_power()}}
}
