% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-ci.R
\name{rate_ci}
\alias{rate_ci}
\title{Estimate Rate with Confidence Limits}
\usage{
rate_ci(r, n)
}
\arguments{
\item{r}{A numeric vector of the number of returns in each group.}

\item{n}{A numeric vector of the number of individuals in each group.
If n is a scalar then it is assumed to be the total number of individuals
split evenly between the groups.}
}
\value{
A data frame with six columns of r, n, the estimate of the rate
and lower and upper 95\% confidence limits.
}
\description{
Estimate Rate with Confidence Limits
}
\examples{
rate_ci(5, 10)
}
