% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-effect.R
\name{rate_effect}
\alias{rate_effect}
\title{Estimate Rates with Confidence Limits and P-Values}
\usage{
rate_effect(r, n, alternative = "two.sided")
}
\arguments{
\item{r}{A numeric vector of the number of individuals returning in each group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}
}
\value{
A data frame with six columns of group, r, n, the estimate of the rate,
the lower and upper 95\% confidence limits and the p-value.
}
\description{
Estimates 95\% confidence limits using Wald method.
The p-value for the first group is for p = 0.5 while the p-value for the
subsequent groups is for equal probabilities.
}
\examples{
rate_effect(5, 10)
rate_effect(c(1,9), c(10,10))
rate_effect(c(1,9), 20)
}
\seealso{
\code{\link[=rate_sim]{rate_sim()}} and \code{\link[=rate2_power_analysis]{rate2_power_analysis()}}
}
