% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-effect-bayesian.R
\name{rate_effect_bayesian}
\alias{rate_effect_bayesian}
\title{Estimate Rates with Credible Limits and P-Values}
\usage{
rate_effect_bayesian(r, n, alpha = 1, beta = 1, alternative = "two.sided")
}
\arguments{
\item{r}{A numeric vector of the number of individuals returning in each group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}

\item{alpha}{A numeric vector of the first shape parameter for the beta distribution.}

\item{beta}{A numeric vector of the second shape parameter for the beta distribution.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}
}
\value{
A data frame with eight columns of group, r, n, alpha, beta, the estimate of the rate,
the lower and upper 95\% confidence limits and the p-value.
}
\description{
Estimates rates with 95\% credible limits.
The credible limits and p-values are estimate using MCMC sample.
The p-value for the first group is for p1 = 0.5 while the p-value for
subsequent groups is for pn = p1.
}
\examples{
\dontrun{
rate_effect_bayesian(0, 5)
rate_effect_bayesian(0, 5, 10, 10)
}
}
\seealso{
\code{\link[=rate_effect]{rate_effect()}}
}
