% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-pvalue.R
\name{rate_pvalue}
\alias{rate_pvalue}
\title{Calculate P-Value for Rate}
\usage{
rate_pvalue(r, n)
}
\arguments{
\item{r}{A numeric vector of the number of individuals returnning in each group.}

\item{n}{The total number of individuals in the study or
if a numeric vector the number individuals in each group.}
}
\value{
A data frame with four columns of group, r, n, and the p-value.
The p-value for the first group is for p = 0.5 while the p-value for the
subsequent groups is for equal probabilities.
}
\description{
Calculate P-Value for Rate
}
\examples{
rate_pvalue(5, 10)
rate_pvalue(c(1,9), c(10,10))
rate_pvalue(c(1,9), 20)
}
