% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-sim.R
\name{rate_sim}
\alias{rate_sim}
\title{Simulate Data for Rates}
\usage{
rate_sim(p, n)
}
\arguments{
\item{p}{A numeric vector of the proportion of individuals expected to return in each group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}
}
\value{
A real vector of the number in the first and second group.
}
\description{
Simulate Data for Rates
}
\examples{
set.seed(99)
rate_sim(0.9, 10)
set.seed(99)
rate_sim(0.9, 10)
rate_sim(0.9, 10)
set.seed(100)
rate_sim(c(0.9, 0.3), 20)
}
