% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFOsamplestats-package.R
\docType{package}
\name{DFOsamplestats-package}
\alias{DFOsamplestats}
\alias{DFOsamplestats-package}
\title{DFOsamplestats: a package for sample statistics and testing data relationships}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/DFOsamplestats}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Colton Stephens \email{coltonstephens@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-8744-6405}{ORCID}) [contributor]
}

}
\keyword{internal}
