% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for functions}
\arguments{
\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{alpha}{A numeric vector of the first shape parameter for the beta distribution.}

\item{beta}{A numeric vector of the second shape parameter for the beta distribution.}

\item{p1}{A numeric scalar of the proportion of individuals expected to return from the first group.}

\item{p2}{A numeric scalar of the proportion of individuals expected to return from the second group.}

\item{p}{A numeric vector of the proportion of individuals expected to return in each group.}

\item{r}{A numeric vector of the number of individuals returning in each group.}

\item{n}{The total number of individuals in the study (divided evenly among groups) or
if a numeric vector the number individuals in each group.}

\item{nsims}{Number of simulations to perform.}
}
\description{
Descriptions of the parameters for functions
}
\keyword{internal}
