% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for functions}
\arguments{
\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{n}{The total number of individuals in the study or
if a numeric vector the number individuals in each group.}

\item{nsims}{Number of simulations to perform.}

\item{p1}{Expected proportion of individuals returning from first group.}

\item{p2}{Expected proportion of individuals returning from second group.}

\item{r}{A numeric vector of the number of individuals returnning in each group.}
}
\description{
Descriptions of the parameters for functions
}
\keyword{internal}
