% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-estimate.R
\name{prop2_estimate}
\alias{prop2_estimate}
\title{Estimate, Confidence Interval and P-Value for Exact Two Sample Test Proportions}
\usage{
prop2_estimate(n1, n2, one_sided = FALSE)
}
\arguments{
\item{n1}{Number of samples in first group.}

\item{n2}{Number of samples in second group.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}
}
\value{
A numeric vector of the estimate and lower and upper 95\% confidence limits.
}
\description{
Uses \code{stats::binom.test()} to calculate the estimate 95\% Confidence Interval
and P-value for the probability of being in n1 vs n2.
}
\examples{
prop2_estimate(6, 0)
prop2_estimate(3, 3)
prop2_estimate(0, 6)
prop2_estimate(106, 100)
prop2_estimate(106, 100, one_sided = TRUE)
prop2_estimate(100, 106, one_sided = TRUE)
}
