% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-prop2-test.R
\name{prop2_power}
\alias{prop2_power}
\title{P-value for Exact Two Sample Test Proportions}
\usage{
prop2_power(n1, n2, one_sided = FALSE)
}
\arguments{
\item{n1}{Number of samples in first group.}

\item{n2}{Number of samples in second group.}

\item{one_sided}{A flag specifying whether to perform a one sided test.}
}
\value{
A numeric scalar of the p-value.
}
\description{
Uses \code{stats::binom.test()} to calculate the exact p-value for a
a two sample test of a difference in the probability of being in n1 vs n2.
}
\examples{
prop2_power(6, 0)
prop2_power(3, 3)
prop2_power(0, 6)
prop2_power(106, 100)
prop2_power(106, 100, one_sided = TRUE)
prop2_power(100, 106, one_sided = TRUE)
}
