% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-power-analysis.R
\name{prop2_power_analysis}
\alias{prop2_power_analysis}
\title{Power Analysis for Two Sample Test Proportion}
\usage{
prop2_power_analysis(p1, n, alternative = "two.sided", nsims = 1000)
}
\arguments{
\item{p1}{Proportion appearing in first group.}

\item{n}{Total number of samples in study. This is different to
\code{stats::power.prop.test()} where \code{n} is the number of samples in each group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{nsims}{Number of simulations to perform.}
}
\value{
A real scalar of the power.
}
\description{
Power Analysis for Two Sample Test Proportion
}
\examples{
set.seed(99)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.9, 10)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.1, 10)
set.seed(99)
prop2_power_analysis(0.45, 500)
}
