% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-power-analysis.R
\name{prop2_power_analysis}
\alias{prop2_power_analysis}
\title{Power Analysis for Two Sample Test Proportion}
\usage{
prop2_power_analysis(p1, n, alternative = "two.sided", nsims = 1000)
}
\arguments{
\item{p1}{Probability appearing in first vs second group.}

\item{n}{Total number of samples.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{nsims}{A whole number of the number of simulations to perform.}
}
\value{
A real scalar of the power.
}
\description{
Power Analysis for Two Sample Test Proportion
}
\examples{
set.seed(99)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.9, 10)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.1, 10)
set.seed(99)
prop2_power_analysis(0.45, 500)
}
