% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-power-analysis.R
\name{prop2_power_analysis}
\alias{prop2_power_analysis}
\title{Power Analysis for Two Sample Test Proportion}
\usage{
prop2_power_analysis(
  p1,
  n,
  p2 = 1 - p1,
  alternative = "two.sided",
  nsims = 1000
)
}
\arguments{
\item{p1}{Expected proportion of individuals returning from first group.}

\item{n}{The total number of individuals in the study or
if a numeric vector the number individuals in each group.}

\item{p2}{Expected proportion of individuals returning from second group.}

\item{alternative}{A string specifying whether to perform a "two.sided",
or "less" or "greater" one-sided test.}

\item{nsims}{Number of simulations to perform.}
}
\value{
A real scalar of the power.
}
\description{
Power Analysis for Two Sample Test Proportion
}
\examples{
set.seed(99)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.9, 10)
prop2_power_analysis(0.9, 10)
set.seed(99)
prop2_power_analysis(0.1, 10)
set.seed(99)
prop2_power_analysis(0.45, 500)
}
