% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2-sim.R
\name{prop2_sim}
\alias{prop2_sim}
\title{Simulate Data for Two Sample Test Proportion}
\usage{
prop2_sim(p1, n, p2 = 1 - p1)
}
\arguments{
\item{p1}{Expected proportion of individuals returning from first group.}

\item{n}{The total number of individuals in the study or
if a numeric vector the number individuals in each group.}

\item{p2}{Expected proportion of individuals returning from second group.}
}
\value{
A real vector of the number in the first and second group.
}
\description{
Simulate Data for Two Sample Test Proportion
}
\examples{
set.seed(99)
prop2_sim(0.9, 10)
set.seed(99)
prop2_sim(0.9, 10)
prop2_sim(0.9, 10)
set.seed(99)
prop2_sim(0.1, 10)
set.seed(99)
prop2_sim(0.45, 10)
}
