% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn-estimate.R
\name{propn_estimate}
\alias{propn_estimate}
\title{P-value for Exact Multiple Sample Test Proportions}
\usage{
propn_estimate(ns)
}
\arguments{
\item{ns}{A vector of the number of samples in each group.}
}
\value{
A numeric scalar of the p-value.
}
\description{
Uses \code{stats::prop.test()} to calculate the exact p-value for a
multiple sample test of a difference in the probability of being in the groups.
}
\examples{
propn_estimate(c(106, 100))
propn_estimate(c(106, 100, 90))
propn_estimate(c(106, 80))
propn_estimate(c(106, 80, 106))
}
