% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn-pvalue.R
\name{propn_pvalue}
\alias{propn_pvalue}
\title{P-value for Multiple Sample Test Proportions}
\usage{
propn_pvalue(ns)
}
\arguments{
\item{ns}{A vector of the number of samples in each group.}
}
\value{
A numeric scalar of the p-value.
}
\description{
P-value for Multiple Sample Test Proportions
}
\examples{
propn_pvalue(c(106, 100))
propn_pvalue(c(106, 100, 90))
propn_pvalue(c(106, 80))
propn_pvalue(c(106, 80, 106))
}
