% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-ci.R
\name{rate_ci}
\alias{rate_ci}
\title{Estimate Rate with Confidence Limits}
\usage{
rate_ci(r, n)
}
\arguments{
\item{r}{A numeric vector of the number of individuals returnning in each group.}

\item{n}{The total number of individuals in the study or
if a numeric vector the number individuals in each group.}
}
\value{
A data frame with six columns of group, r, n, the estimate of the rate
and lower and upper 95\% confidence limits.
}
\description{
Estimates 95\% confidence limits using Wald method.
}
\examples{
rate_ci(5, 10)
}
